Create README.md
#Hepsiburada-Web Scraper Project

The provided code scrapes data from a website using BeautifulSoup and requests library, extracts the title and price of an iPhone 14 128 GB, cleans up the price data and stores it in a CSV file. It then combines all the code into a function "check_price" which can be called to update the data in the CSV file. The while loop at the end will run the function continuously after a set time and input the data into the CSV file.

#Movie Correlation Project

In this project, I did correlation analysis using the movie data available on Kaggle(which was scraped from IMDB). I used four essential libraries, namely Pandas, Numpy, Matplotlib, and Seaborn, I was able to efficiently read in, clean, and prepare the data for analysis. I identified and handled any missing data to ensure accurate results. With the data thoroughly preprocessed, I used various techniques to uncover correlations within the dataset.